fix: keep worker diagnostics within package file budget
